[UNI-209] refactor : node 도메인 -> map 도메인 이동
