+++ uniro_backend/src/main/java/com/softeer5/uniro_backend/map/repository/NodeRepository.java
@@ -1,4 +1,4 @@
-package com.softeer5.uniro_backend.building.repository;
+package com.softeer5.uniro_backend.map.repository;
 
 import com.softeer5.uniro_backend.map.entity.Node;
 import org.springframework.data.jpa.repository.JpaRepository;
